Primary stuff

+++ reviews/router.py
@@ -1,37 +1,39 @@
-from datetime import datetime
-from typing import Dict
-
-from fastapi import APIRouter, Body, FastAPI, HTTPException
-from fastapi.responses import JSONResponse
+import os
+from typing import Optional
+from loguru import logger
+from dotenv import load_dotenv
+from fastapi import APIRouter, FastAPI
 from starlette import status
 
-from customer.schemas import CustomerUpdateResponse
+from reviews.models import ReviewTable
 from reviews.schemas import (DeleteReviewResponse, PostReviewRequest,
                              PostReviewResponse, PutReviewRequest,
-                             PutReviewResponse, Review)
+                             PutReviewResponse, Review, GetReviewsResponse)
 
 # Define router
 router = APIRouter(prefix="/reviews", tags=["Reviews Management"])
 
-# reviews: dict[tuple[str, str], Review] = {}
+load_dotenv()
 
-reviews: dict[str, dict[str, Review]] = {}
+db: ReviewTable = ReviewTable(url=os.getenv("SUPABASE_URL"), key=os.getenv("SUPABASE_KEY"))
 
 
 @router.post("/submit")
 async def submit_review(review: PostReviewRequest):
     try:
 
-        key1, key2 = review.item_id, review.customer_id
-        if key1 not in reviews or key2 not in reviews[key1]:
+        if db.get_review(item_id=review.item_id, customer_id=review.customer_id):
             return PostReviewResponse(
                 status_code=status.HTTP_409_CONFLICT, review_schema=review
             )
-
-        reviews[key1][key2] = Review(
-            **review.model_dump(), time=datetime.now(), flagged="needs_approval"
+        new_review: Review = Review(
+            **review.model_dump(), flagged="needs_approval"
         )
-        return PostReviewResponse(status_code=status.HTTP_200_OK, review_schema=review)
+        if db.submit_review(review=new_review):
+            return PostReviewResponse(status_code=status.HTTP_200_OK, review_schema=review)
+
+        return PostReviewResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, review_schema=review,
+                                  errors="Could not submit review")
 
     except Exception as e:
         return PostReviewResponse(
@@ -44,9 +46,14 @@ async def submit_review(review: PostReviewRequest):
 @router.put("/update")
 async def update_review(updated_review: PutReviewRequest):
     try:
-        key: tuple[str, str] = (updated_review.item_id, updated_review.customer_id)
-        stored_review = reviews[key]
-        updated_data = stored_review.model_dump()
+        stored_review: Optional[list[Review]] = \
+            db.get_review(item_id=updated_review.item_id, customer_id=updated_review.customer_id)
+        if not stored_review:
+            return PutReviewResponse(
+                status_code=status.HTTP_409_CONFLICT, review_schema=updated_review
+            )
+
+        updated_data = stored_review[0].model_dump()
         if updated_review.model_dump(exclude_unset=True).items():
             for _key, value in updated_review.model_dump(exclude_unset=True).items():
                 updated_data[_key] = value
@@ -54,6 +61,14 @@ async def update_review(updated_review: PutReviewRequest):
             return PutReviewResponse(
                 status_code=status.HTTP_202_ACCEPTED, review_schema=updated_review
             )
+
+        if not db.update_review(review=Review.model_validate(updated_data)):
+            return PutReviewResponse(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                review_schema=updated_review,
+                errors="Failed to update review",
+            )
+
         return PutReviewResponse(
             status_code=status.HTTP_200_OK,
             review_schema=updated_review,
@@ -68,19 +83,23 @@ async def update_review(updated_review: PutReviewRequest):
 
 
 @router.delete("/delete")
-async def delete_review(customer_id: str, item_id: str):
+async def delete_review(customer_id: str, item_id: int):
     try:
-        key: tuple[str, str] = (item_id, customer_id)
-        if key not in reviews:
+
+        if not db.get_review(item_id=item_id, customer_id=customer_id):
             return DeleteReviewResponse(
                 status_code=status.HTTP_404_NOT_FOUND,
                 customer_id=customer_id,
                 item_id=item_id,
             )
-        del reviews[key]
-        return DeleteReviewResponse(
-            status_code=status.HTTP_200_OK, customer_id=customer_id, item_id=item_id
-        )
+        # if not db
+        if db.delete_review(item_id=item_id, customer_id=customer_id):
+            return DeleteReviewResponse(
+                status_code=status.HTTP_200_OK, customer_id=customer_id, item_id=item_id
+            )
+
+        raise Exception("Failed to delete review. Database Error.")
+
     except Exception as e:
         return DeleteReviewResponse(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -90,27 +109,52 @@ async def delete_review(customer_id: str, item_id: str):
         )
 
 
-@router.get("/product/{product_id}")
-async def get_review(product_id: str):
+async def get_generic_review(item_id: Optional[int] = None, customer_id: Optional[str] = None) -> GetReviewsResponse:
+    assert item_id or customer_id, "Invalid Call"
+    try:
+        filters: dict = {}
+        if item_id:
+            filters["item_id"] = item_id
+        if customer_id:
+            filters["customer_id"] = customer_id
+
+        reviews: Optional[list[Review]] = db.get_reviews(**filters)
+        if reviews is None:
+            return GetReviewsResponse(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, reviews=reviews, item_id=item_id,
+                customer_id=customer_id,
+                errors="Database Error"
+            )
+        return GetReviewsResponse(
+            status_code=status.HTTP_200_OK, reviews=reviews, item_id=item_id, customer_id=customer_id,
+        )
+    except Exception as e:
+        logger.exception(e)
+        return GetReviewsResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, item_id=item_id, errors=str(e)
+        )
+
+
+@router.get("/item/{item_id}")
+async def get_item_review(item_id: int):
     # Get From Database
-    pass
+    return get_generic_review(item_id=item_id)
 
 
 @router.get("/customer/{customer_id}")
 async def get_customer_review(customer_id: str):
-    # Get From Database
-    pass
+    return get_generic_review(customer_id=customer_id)
+
+
+@router.get("/details")
+async def get_details(customer_id: str, item_id: int):
+    # Get from database
+    return get_generic_review(item_id=item_id, customer_id=customer_id)
 
 
 @router.put("/moderate")
 async def moderate():
     # Update Database
-    pass
-
-
-@router.get("/details")
-async def get_details():
-    # Get from database
     pass
